[Rename] 프로필 이미지 관련 컴포넌트 폴더분리
