+++ client/src/components/ProfileImage/index.js
@@ -1,0 +1,5 @@
+import ProfileImg from './ProfileImg';
+import ProfileImgSetter from './ProfileImgSetter';
+import ProfileUploader from './ProfileUploader';
+
+export { ProfileImg, ProfileImgSetter, ProfileUploader };
